+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Version 1.4.4 (2023-11-27)
 
+- added date of last update
 - added professional activities to output 
 - added type of supervision
 - updated on qualis table for 2017-2020 and SJR for 2022